backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
pymongo
requests
motor
python-multipart
numpy
pillow
torch==2.4.0
torchvision==0.19.0
keras==3.3.3
transformers
minio
trino
polars
confluent_kafka
kafka-python
